+++ backend/src/main/java/com/example/demo/service/TeamDiaryServiceImpl.java
@@ -27,26 +27,26 @@ public class TeamDiaryServiceImpl implements TeamDiaryService {
     }
 
     // 팀 일기 수정
-    public void updateTeamDiary(int teamDiaryId, TeamDiaryModel teamDiary) {
+    public void updateTeamDiary(long teamDiaryId, TeamDiaryModel teamDiary) {
         teamDiary.setId(teamDiaryId);
         TeamDiaryMapper.updateTeamDiary(teamDiary);
     }
 
     // 팀 일기 삭제 (공유 해제)
     @Override
-    public void deleteTeamDiary(int diaryId, int teamId) {
+    public void deleteTeamDiary(long diaryId, long teamId) {
         TeamDiaryMapper.deleteTeamDiary(diaryId, teamId);
     }
 
     // 현재 팀에 공유된 일기 리스트 요청
     @Override
-    public List<TeamDiaryListResponse> requestTeamDiaryList(int teamId) {
+    public List<TeamDiaryListResponse> requestTeamDiaryList(long teamId) {
         return TeamDiaryMapper.requestTeamDiaryList(teamId);
     }
 
     // 선택한 일기가 공유된 팀들의 id, 이름 요청
     @Override
-    public List<SharedTeamsResponse> requestSharedTeams(int diaryId) {
+    public List<SharedTeamsResponse> requestSharedTeams(long diaryId) {
         return TeamDiaryMapper.requestSharedTeams(diaryId);
     }
 }
